[ME] Separate PluginBase configuration, parsing and asset helpers
