+++ src/MaterialEditor.Base/Assets/MaterialTextureExport.cs
@@ -1,0 +1,43 @@
+using System.IO;
+using UnityEngine;
+
+namespace MaterialEditorAPI
+{
+    internal static class MaterialTextureExport
+    {
+        internal static Texture2D GetT2D(RenderTexture renderTexture)
+        {
+            var currentActiveRT = RenderTexture.active;
+            RenderTexture.active = renderTexture;
+            var tex = new Texture2D(renderTexture.width, renderTexture.height);
+            tex.ReadPixels(new Rect(0, 0, tex.width, tex.height), 0, 0);
+            RenderTexture.active = currentActiveRT;
+            return tex;
+        }
+
+        internal static void SaveTexR(RenderTexture renderTexture, string path)
+        {
+            var tex = GetT2D(renderTexture);
+            File.WriteAllBytes(path, EncodeTextureToPng(tex));
+            Object.DestroyImmediate(tex);
+        }
+
+        internal static byte[] EncodeTextureToPng(Texture2D texture)
+        {
+            return texture.EncodeToPNG();
+        }
+
+        internal static void SaveTex(Texture tex, string path, RenderTextureFormat rtf = RenderTextureFormat.Default, RenderTextureReadWrite cs = RenderTextureReadWrite.Default)
+        {
+            var tmp = RenderTexture.GetTemporary(tex.width, tex.height, 0, rtf, cs);
+            var currentActiveRT = RenderTexture.active;
+            RenderTexture.active = tmp;
+            GL.Clear(false, true, new Color(0, 0, 0, 0));
+            Graphics.Blit(tex, tmp);
+            SaveTexR(tmp, path);
+            RenderTexture.active = currentActiveRT;
+            RenderTexture.ReleaseTemporary(tmp);
+        }
+
+    }
+}
